Switch to conda; attempt to use pysftp
pysftp, which is no longer developed, does not appear compatible
with current versions of paramiko, judging by the exception shown
in the notebook (precompute.ipynb)

Assuming that's the case, I don't want to use an old version of
paramiko, in part because the convenience of using pysftp is not
great enough to justify the hassle of using an older version, but
more importantly because I'd need to look into the issue of any
security vulnerabilities patched between the old version I would
use and the current version.

So I'll probably switch from pysftp to fabric (or possibly use
paramiko directly).