Start investigating nondeterminism
Embeddings seem nondeterministic, but I'm not yet certain I haven't
made a mistake in serialization that is causing it or contributing.

The extra sections added in embeddings.ipynb to investiate this are
kind of a mess, and should probably be revise or (if they don't
provide lasting value) removed when the investigation is done.