Fix dev container carry-in
This corresponds to https://github.com/EliahKagan/palgoviz/pull/58.

I had very closely based the configuration here on the palgoviz
configuration, including copying and pasting some parts. As
detailed in that PR, I had tried to test all combinations/platforms
for palgoviz, but missed the one where the situation we need to
carry user.name and user.email in actually tends to occur. Here, I
had not attempted the kind of extensive manual testing that I had
(erroneously) thought I had done there, so my mistake there carried
over to this repository.

This fixes the bug by adding the missing logic in postCreate to
apply the configuration, if any, that initialize or initialize.cmd
obtained.