Don't list indirect openai dependencies (conda)
This removes dependencies from environment.yml that are indirect
dependencies, via openai[embeddings]. In a previous version of the
openai conda-forge package, I believe the indirect dependencies
provided by the "embeddings" extra were not include. But now they
are, so they need not be listed in environmeny.yml.

The openai package's dependencies are subject to change, and I've
stilled listed packages (e.g., pandas) that I am depending on for
reasons beyond being required to successfully import the
openai.embeddings_utils module.

This is the conda analogue of the previous commit (which applies to
PyPI/poetry dependencies).

(I've made such the notebooks still run with this change, but I am
not committing those slight changes in the notebooks.)