Update interface in moderation.py
This adds the new categories to the TypedDict definitions in the
moderation submodule.

It also changes the Result.flagged docstring because I'm not
confident my old docstring was (and continues to be correct), and
adds a convenience function to do what I had previously documented
Result.flagged as having done.

I've also fixed a bug where __all__ didn't include get_moderation.