Chunk moderation requests according to new limits
This sends multiple requests, each of up to 32 texts, rather than
sending a single request with all 700+ texts. That had worked
until recently, but the error message indicates a limit of 32.

(This also fixes the call to a new moderation-related function in
the main notebook, which needed to be qualified with module names.)

The run was successful, and new data are accordingly committed.