Update dependencies, including cryptography
This updates direct and indirect dependencies, including the update
of the cryptography package (which is a dependency of paramiko,
which is a dependency of fetch) from 41.0.1 to 41.0.2, patching
CVE-2023-38325.

This is only for poetry. I am not pinning versions for conda/mamba,
and since cryptography is an indirect dependency, it is not listed
in environment.yml. At this time, I have not added cryptography to
environment.yml for the purpose of specifying >=41.0.2. This is
unlike with requests, where I had long meant to do that and
recently did; requests is a direct dependency and was already
listed. However, this is like with tornado (another indirect
dependency that had an important CVE that was patched).