Rerun notebooks
This removes stale saved embeddings and moderation model output,
as usual.

Rerunning pricing.ipynb removed all rows from the able except the
one for Ada v2, but that's the one this project needs, so that's
fine. This happened because OpenAI has moved the pricing
information for their first-generation embedding models to:
https://platform.openai.com/docs/deprecations/first-generation-text-embedding-models