Update dependencies (including aiohttp)
And re-run notebooks to test.

The update to the indirect aiohttp dependency takes it from 3.8.4
to 3.8.5, patching CVE-2023-37276 (which is due to CVE-2023-30589).
Because this project's indirect dependency on aiohttp is only
through the openai package, which uses it as an HTTP client and is
unlikely to begin using it as an HTTP server as well, this project
is not really affected, so this update to findrepo2-experiment
might be regarded as a regular dependency update rather than a
security update.